+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `AzureSignToolUtils` sign to ignore exception.
 - Update `AzureSignToolUtils` to ignore if file is already signed.
 - Update `HasSignature` to use `PathTooLongUtils` to check if file is signed. (Fix: #77)
+- Update `ExecuteReleaseAsset` to execute `IHazAssetRelease` and build with `IAssetRelease`
 ### Example
 - Add `Resource` and `Resource.pt-BR` to test sign files.
 ### Tests